Move API methods and interfaces to lib
