+++ lib/blocks-api/mapProperties.ts
