+++ pages/api/v1/databases/[id]/index.ts
@@ -5,10 +5,10 @@ import { getSpaceFromApiKey, onError, onNoMatch, requireApiKey } from 'lib/middl
 import { filterObjectKeys } from 'lib/utilities/objects';
 import { NextApiRequest, NextApiResponse } from 'next';
 import nc from 'next-connect';
-import { CardFromBlock } from 'pages/api/v1/databases/card.class';
+import { CardFromBlock } from 'lib/blocks-api/card.class';
 import { validate } from 'uuid';
-import { BoardPage, CardProperty, CardQuery, PaginatedQuery, PaginatedResponse, Card } from '../interfaces';
-import { mapProperties } from './mapProperties';
+import { BoardPage, CardProperty, CardQuery, PaginatedQuery, PaginatedResponse, Card } from 'lib/blocks-api/interfaces';
+import { mapProperties } from 'lib/blocks-api/mapProperties';
 
 const handler = nc<NextApiRequest, NextApiResponse>({ onError, onNoMatch });
